When registering don't send project as empty string
